Se realizan mejoras en los controllers para ajustarse a la architectura final usada en clase, se agrega el meta, httpresponse y Global Handler

+++ src/main/java/com/project/demo/rest/category/CategoryRestController.java
@@ -2,12 +2,22 @@ package com.project.demo.rest.category;
 
 import com.project.demo.logic.entity.category.Category;
 import com.project.demo.logic.entity.category.CategoryRepository;
+import com.project.demo.logic.entity.http.GlobalResponseHandler;
+import com.project.demo.logic.entity.http.Meta;
 import com.project.demo.logic.entity.product.Product;
+import com.project.demo.logic.entity.user.User;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/categories")
@@ -16,14 +26,31 @@ public class CategoryRestController {
     private CategoryRepository categoryRepository;
 
     @GetMapping
-    public List<Category> getAllCategories() {
-        return (List<Category>) categoryRepository.findAll();
+    @PreAuthorize("hasAnyRole('SUPER_ADMIN_ROLE')")
+    public ResponseEntity<?> getAll(
+            @RequestParam(defaultValue = "1") int page,
+            @RequestParam(defaultValue = "10") int size,
+            HttpServletRequest request) {
+
+        Pageable pageable = PageRequest.of(page-1, size);
+        Page<Category> categoriesPage = categoryRepository.findAll(pageable);
+        Meta meta = new Meta(request.getMethod(), request.getRequestURL().toString());
+        meta.setTotalPages(categoriesPage.getTotalPages());
+        meta.setTotalElements(categoriesPage.getTotalElements());
+        meta.setPageNumber(categoriesPage.getNumber() + 1);
+        meta.setPageSize(categoriesPage.getSize());
+
+        return new GlobalResponseHandler().handleResponse("Order retrieved successfully",
+                categoriesPage.getContent(), HttpStatus.OK, meta);
     }
 
     @PostMapping
     @PreAuthorize("hasAnyRole('SUPER_ADMIN_ROLE')")
-    public Category addCategory(@RequestBody Category category) {
-        return categoryRepository.save(category);
+    public ResponseEntity<?> addCategory(@RequestBody Category category, HttpServletRequest request) {
+
+        categoryRepository.save(category);
+        return new GlobalResponseHandler().handleResponse("Category created successfully",
+                category, HttpStatus.OK, request);
     }
 
     @GetMapping("/{id}")
@@ -33,24 +60,32 @@ public class CategoryRestController {
 
     @PutMapping("/{id}")
     @PreAuthorize("hasAnyRole('SUPER_ADMIN_ROLE')")
-    public Category updateCategory(@PathVariable Long id, @RequestBody Category category) {
-        return categoryRepository.findById(id)
-                .map(existingCategory -> {
-                    existingCategory.setName(category.getName());
-                    existingCategory.setDescription(category.getDescription());
-
-                    return categoryRepository.save(existingCategory);
-                })
-                .orElseGet(() -> {
-                    category.setId(id);
-                    return categoryRepository.save(category);
-                });
+    public ResponseEntity<?> updateCategory(@PathVariable Long id, @RequestBody Category category, HttpServletRequest request) {
+        Optional<Category> foundCategory = categoryRepository.findById(id);
+        if(foundCategory.isPresent()) {
+            category.setId(id);
+            category.setProducts(foundCategory.get().getProducts());
+            categoryRepository.save(category);
+            return new GlobalResponseHandler().handleResponse("Category updated successfully",
+                    category, HttpStatus.OK, request);
+        } else {
+            return new GlobalResponseHandler().handleResponse("Category id " + id + " not found"  ,
+                    HttpStatus.NOT_FOUND, request);
+        }
     }
 
     @DeleteMapping("/{id}")
     @PreAuthorize("hasAnyRole('SUPER_ADMIN_ROLE')")
-    public void deleteCategory(@PathVariable Long id) {
-        categoryRepository.deleteById(id);
+    public ResponseEntity<?> deletecategory(@PathVariable Long id, HttpServletRequest request) {
+        Optional<Category> foundCategory = categoryRepository.findById(id);
+        if(foundCategory.isPresent()) {
+            categoryRepository.deleteById(id);
+            return new GlobalResponseHandler().handleResponse("Category deleted successfully",
+                    foundCategory.get(), HttpStatus.OK, request);
+        } else {
+            return new GlobalResponseHandler().handleResponse("Category id " + id + " not found"  ,
+                    HttpStatus.NOT_FOUND, request);
+        }
     }
 
 }
